Osu-NetApi changes. Done userUrl detect

+++ WAV-Osu-NetApi/Bancho/Models/UserStatistics.cs
@@ -9,12 +9,12 @@ namespace WAV_Osu_NetApi.Bancho.Models
         public UserLevel level { get; set; }
         public int global_rank { get; set; }
         public double pp { get; set; }
-        public int ranked_score { get; set; }
+        public long ranked_score { get; set; }
         public double hit_accuracy { get; set; }
         public int play_count { get; set; }
         public int play_time { get; set; }
-        public int total_score { get; set; }
-        public int total_hits { get; set; }
+        public long total_score { get; set; }
+        public long total_hits { get; set; }
         public int maximum_combo { get; set; }
         public int replays_watched_by_others { get; set; }
         public bool is_ranked { get; set; }
